day8/requirements.txt
asgiref==3.11.1
click==8.3.2
croniter==6.2.2
Django==4.2.29
django-constance==4.3.5
django-debug-toolbar==6.1.0
django-redis==6.0.0
django-rq==4.1.0
django-simple-history==3.10.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
packaging==26.0
psycopg==3.3.3
psycopg-binary==3.3.3
psycopg2-binary==2.9.11
PyJWT==2.12.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
redis==7.0.1
rq==2.7.0
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
